requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
nltk==3.4.4
numpy==1.16.4
pandas==0.25.0
praw==6.3.1
prawcore==1.0.1
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.5
update-checker==0.16
urllib3==1.25.3
websocket-client==0.56.0
Werkzeug==0.15.5
